fix(materialize): let @s<n> govern primitive width, not the range bounds
Move size derivation onto TypeDecl as sizeBytes/sizeBits, add a generic
children accessor, and fix three width bugs it exposed.

Cursor.readRangeBound truncates range bounds to the low 64 bits, so gcc's
unsigned maxima arrive negative — `unsigned long long`'s max is emitted as
01777777777777777777777 and lands as -1L. Width classification must compare
unsigned or it reads that as 1 byte. Signed bounds also keyed off -0x100 /
-0x10000 / -0x100000000 rather than the true minima, widening every subrange
by a factor of two.

The ordering bug is the older one: WithSizeAttr resolved its inner descriptor
first and only consulted @s<n> as a fallback, so the inner's own bounds always
won. gcc emits @s<n> precisely where the bounds cannot carry the answer —
`@s128;r(0,25);0;0377…;` truncates to 0..-1, which every gcc 3.4.5 binary in
the corpus was classifying as an 8-byte ulonglong. Width now comes from the
attribute and identity from the inner, so a slot still names its primitive and
a char range stays char (@s8;r(0,10);-128;127;). resolveSizedRange is keyed on
bits rather than bytes, so an odd @s<n> falls through instead of rounding up
into a plausible byte width, and 128 maps to UnsignedInteger16DataType —
signedness is unrecoverable there, both signs truncate alike, but the width is
what layout depends on.

Sizes stay Long: they are untrusted values that get arithmetic done on them
(sizeBytes * 8, offsetBits + sizeBits) and compared against FieldDecl's
already-Long widths, and usefulStructSize / the gcc-12 detector exist because
stabs routinely overstate them. Pointer and Reference report null rather than
guessing an address size the stab never states.

+++ src/main/kotlin/ghistabs/parse/Parser.kt
@@ -26,7 +26,7 @@ class Parser(src: String, sink: DiagnosticSink = DummySink) : DiagnosticSink by 
 
         // gcc's XCOFF builtin slot for bool, and the shape its `-gstabs+` spelling decodes to.
         const val BUILTIN_BOOL = -16
-        const val BITS_PER_BYTE = 8
+        const val BITS_PER_BYTE = 8L
         val BOOL_ENUM_MEMBERS = listOf("False" to 0L, "True" to 1L)
 
         // Symbol chars that may follow `operator` in a method name (arithmetic, logical,
@@ -581,7 +581,7 @@ class Parser(src: String, sink: DiagnosticSink = DummySink) : DiagnosticSink by 
         val max = readRangeBound()
         consume(';')
         if (max == 0L && min > 0L) {
-            return TypeDecl.Float(min.toInt())
+            return TypeDecl.Float(min)
         }
         return TypeDecl.Range(typeId, min, max)
     }
@@ -596,7 +596,7 @@ class Parser(src: String, sink: DiagnosticSink = DummySink) : DiagnosticSink by 
         consume('R')
         val rCode = readInt().toInt()
         consume(';')
-        val sizeBytes = readInt().toInt()
+        val sizeBytes = readInt()
         consume(';')
         consume('0')
         consume(';')
@@ -632,7 +632,7 @@ class Parser(src: String, sink: DiagnosticSink = DummySink) : DiagnosticSink by 
     private fun Cursor.parseSizeAttr(): TypeDecl.WithSizeAttr<LocalTypeId> {
         consume('@')
         consume('s')
-        val sizeBits = readInt().toInt()
+        val sizeBits = readInt()
         consume(';')
         val inner = parseType()
         return TypeDecl.WithSizeAttr(sizeBits, inner)
